NavTop adicionado
