+++ my-app/src/NavTop.css
@@ -1,0 +1,83 @@
+.nav-top {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px 135px;
+  gap: 10px;
+  position: relative;
+  width: auto;
+  height: auto;
+  left: 0px;
+  top: 0px;
+  background: #FFFFFF;
+  box-shadow: 0px 4px 20px rgba(0, 0, 0, 0.15);
+  }
+    .nav-infos {
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      padding: 0px;
+      gap: 855px;
+      width: 1170px;
+      height: 45px;    
+    }
+      .logos {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        padding: 0px;
+        gap: 20px;
+        width: 292px;
+        height: 45px;    
+      }
+        .logo-beeteller {
+          display: flex;
+          gap: 2.35px;
+          width: 120px;
+          height: 21px;     
+        }
+        .divider {
+          width: 1px;
+          height: 45px;      
+          background: #BDBDBD;     
+        }
+        .cotacoes {
+          display: flex;
+          flex-direction: row;
+          align-items: flex-start;
+          padding: 0px;
+          gap: 8px;        
+          width: 131px;
+          height: 24px;    
+        }
+          .t-cotacoes {
+            width: 99px;
+            height: 21px;      
+            font-family: 'Roboto';
+            font-style: normal;
+            font-weight: 500;
+            font-size: 18px;
+            line-height: 21px;
+            letter-spacing: 0.15em;
+            font-variant: all-small-caps;  
+            color: #BDBDBD;   
+          }
+          .activity {
+            width: 24px;
+            height: 24px;   
+          }
+            .v-activity {
+              width: 12px;
+              height: 18px;
+            }
+      .idioma {
+        width: 23px;
+        height: 21px;
+        font-family: 'Roboto';
+        font-style: normal;
+        font-weight: 500;
+        font-size: 18px;
+        line-height: 21px;
+        letter-spacing: -0.04em;
+        color: #333333;     
+      }
